0.2.0 Refactor into modules
