+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "echo-ia"
-version = "0.1.5"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
